StepTwoForm refactored and correspondent code from FormContent removed. "locationType" and its setter moved to useContext instead of useState for purpuses of multi-step form (to keep the "back" button related to set location.
